+++ doc/sphinxext/gen_rst.py
@@ -19,7 +19,7 @@ import sys
 import gzip
 import posixpath
 import subprocess
-from textwrap import dedent
+import warnings
 
 
 # Try Python 2 first, otherwise load from Python 3
@@ -453,9 +453,10 @@ def extract_docstring(filename, ignore_heading=False):
                         first_par = ((first_par[:95] + '...')
                                      if len(first_par) > 95 else first_par)
                     else:
-                        raise ValueError("Docstring not found by gallery",
-                                         "Please check your example's layout",
-                                         " and make sure it's correct")
+                        raise ValueError("Docstring not found by gallery.\n"
+                                         "Please check the layout of your"
+                                         " example file:\n {}\n and make sure"
+                                         " it's correct".format(filename))
                 else:
                     first_par = paragraphs[0]
 
@@ -513,9 +514,9 @@ Examples
     # better than nested.
     seen_backrefs = set()
     generate_dir_rst('.', fhindex, example_dir, root_dir, plot_gallery, seen_backrefs)
-    for dir in sorted(os.listdir(example_dir)):
-        if os.path.isdir(os.path.join(example_dir, dir)):
-            generate_dir_rst(dir, fhindex, example_dir, root_dir, plot_gallery, seen_backrefs)
+    for directory in sorted(os.listdir(example_dir)):
+        if os.path.isdir(os.path.join(example_dir, directory)):
+            generate_dir_rst(directory, fhindex, example_dir, root_dir, plot_gallery, seen_backrefs)
     fhindex.flush()
 
 
@@ -535,7 +536,7 @@ def extract_line_count(filename, target_dir):
         tok_type = token.tok_name[tok_type]
         if tok_type in ('NEWLINE', 'COMMENT', 'NL', 'INDENT', 'DEDENT'):
             continue
-        elif ((tok_type == 'STRING') and check_docstring):
+        elif (tok_type == 'STRING') and check_docstring:
             erow_docstring = erow
             check_docstring = False
     return erow_docstring+1+start_row, erow+1+start_row
@@ -597,12 +598,12 @@ def _thumbnail_div(subdir, full_dir, fname, snippet):
     return ''.join(out)
 
 
-def generate_dir_rst(dir, fhindex, example_dir, root_dir, plot_gallery, seen_backrefs):
+def generate_dir_rst(directory, fhindex, example_dir, root_dir, plot_gallery, seen_backrefs):
     """ Generate the rst file for an example directory.
     """
-    if not dir == '.':
-        target_dir = os.path.join(root_dir, dir)
-        src_dir = os.path.join(example_dir, dir)
+    if not directory == '.':
+        target_dir = os.path.join(root_dir, directory)
+        src_dir = os.path.join(example_dir, directory)
     else:
         target_dir = root_dir
         src_dir = example_dir
@@ -621,14 +622,14 @@ def generate_dir_rst(dir, fhindex, example_dir, root_dir, plot_gallery, seen_bac
         os.makedirs(target_dir)
     sorted_listdir = line_count_sort(os.listdir(src_dir),
                                      src_dir)
-    if not os.path.exists(os.path.join(dir, 'images', 'thumb')):
-        os.makedirs(os.path.join(dir, 'images', 'thumb'))
+    if not os.path.exists(os.path.join(directory, 'images', 'thumb')):
+        os.makedirs(os.path.join(directory, 'images', 'thumb'))
     for fname in sorted_listdir:
         if fname.endswith('py'):
             backrefs = generate_file_rst(fname, target_dir, src_dir, root_dir, plot_gallery)
             new_fname = os.path.join(src_dir, fname)
             _, snippet, _ = extract_docstring(new_fname, True)
-            fhindex.write(_thumbnail_div(dir, dir, fname, snippet))
+            fhindex.write(_thumbnail_div(directory, directory, fname, snippet))
             fhindex.write("""
 
 .. toctree::
@@ -636,7 +637,7 @@ def generate_dir_rst(dir, fhindex, example_dir, root_dir, plot_gallery, seen_bac
 
    %s/%s
 
-""" % (dir, fname[:-3]))
+""" % (directory, fname[:-3]))
             for backref in backrefs:
                 include_path = os.path.join(root_dir, '../modules/generated/%s.examples' % backref)
                 seen = backref in seen_backrefs
@@ -648,8 +649,8 @@ def generate_dir_rst(dir, fhindex, example_dir, root_dir, plot_gallery, seen_bac
                         print('-----------------%s--' % ('-' * len(backref)),
                               file=ex_file)
                         print(file=ex_file)
-                    rel_dir = os.path.join('../../auto_examples', dir)
-                    ex_file.write(_thumbnail_div(dir, rel_dir, fname, snippet))
+                    rel_dir = os.path.join('../../auto_examples', directory)
+                    ex_file.write(_thumbnail_div(directory, rel_dir, fname, snippet))
                     seen_backrefs.add(backref)
     fhindex.write("""
 .. raw:: html
@@ -828,8 +829,6 @@ def generate_file_rst(fname, target_dir, src_dir, root_dir, plot_gallery):
                              'time_%s.txt' % base_image_name)
     thumb_file = os.path.join(thumb_dir, fname[:-3] + '.png')
     time_elapsed = 0
-    time_m = 0
-    time_s = 0
     if plot_gallery and fname.startswith('plot'):
         # generate the plot as png image if file name
         # starts with plot and if it is more recent than an
@@ -968,90 +967,96 @@ def generate_file_rst(fname, target_dir, src_dir, root_dir, plot_gallery):
 
 def embed_code_links(app, exception):
     """Embed hyperlinks to documentation into example code"""
-    try:
-        if exception is not None:
-            return
-        print('Embedding documentation hyperlinks in examples..')
+    if exception is not None:
+        return
+    print('Embedding documentation hyperlinks in examples..')
 
-        # Add resolvers for the packages for which we want to show links
-        doc_resolvers = {}
-        doc_resolvers['sklearn'] = SphinxDocLinkResolver(app.builder.outdir,
-                                                         relative=True)
+    # Add resolvers for the packages for which we want to show links
+    doc_resolvers = {}
+    doc_resolvers['sklearn'] = SphinxDocLinkResolver(app.builder.outdir,
+                                                     relative=True)
 
-        doc_resolvers['matplotlib'] = SphinxDocLinkResolver(
-            'http://matplotlib.org')
+    resolver_urls = {
+        'matplotlib': 'http://matplotlib.org',
+        'numpy': 'http://docs.scipy.org/doc/numpy-1.6.0',
+        'scipy': 'http://docs.scipy.org/doc/scipy-0.11.0/reference',
+    }
+    for this_module, url in resolver_urls.items():
+        try:
+            doc_resolvers[this_module] = SphinxDocLinkResolver(url)
+        except HTTPError as e:
+            print("The following HTTP Error has occurred:\n")
+            print(e.code)
+        except URLError as e:
+            print("\n...\n"
+                  "Warning: Embedding the documentation hyperlinks requires "
+                  "internet access.\nPlease check your network connection.\n"
+                  "Unable to continue embedding `{0}` links due to a URL "
+                  "Error:\n".format(this_module))
+            print(e.args)
 
-        doc_resolvers['numpy'] = SphinxDocLinkResolver(
-            'http://docs.scipy.org/doc/numpy-1.6.0')
+    example_dir = os.path.join(app.builder.srcdir, 'auto_examples')
+    html_example_dir = os.path.abspath(os.path.join(app.builder.outdir,
+                                                    'auto_examples'))
 
-        doc_resolvers['scipy'] = SphinxDocLinkResolver(
-            'http://docs.scipy.org/doc/scipy-0.11.0/reference')
+    # patterns for replacement
+    link_pattern = '<a href="%s">%s</a>'
+    orig_pattern = '<span class="n">%s</span>'
+    period = '<span class="o">.</span>'
 
-        example_dir = os.path.join(app.builder.srcdir, 'auto_examples')
-        html_example_dir = os.path.abspath(os.path.join(app.builder.outdir,
-                                                        'auto_examples'))
+    for dirpath, _, filenames in os.walk(html_example_dir):
+        for fname in filenames:
+            print('\tprocessing: %s' % fname)
+            full_fname = os.path.join(html_example_dir, dirpath, fname)
+            subpath = dirpath[len(html_example_dir) + 1:]
+            pickle_fname = os.path.join(example_dir, subpath,
+                                        fname[:-5] + '_codeobj.pickle')
 
-        # patterns for replacement
-        link_pattern = '<a href="%s">%s</a>'
-        orig_pattern = '<span class="n">%s</span>'
-        period = '<span class="o">.</span>'
+            if os.path.exists(pickle_fname):
+                # we have a pickle file with the objects to embed links for
+                with open(pickle_fname, 'rb') as fid:
+                    example_code_obj = pickle.load(fid)
+                fid.close()
+                str_repl = {}
+                # generate replacement strings with the links
+                for name, cobj in example_code_obj.items():
+                    this_module = cobj['module'].split('.')[0]
 
-        for dirpath, _, filenames in os.walk(html_example_dir):
-            for fname in filenames:
-                print('\tprocessing: %s' % fname)
-                full_fname = os.path.join(html_example_dir, dirpath, fname)
-                subpath = dirpath[len(html_example_dir) + 1:]
-                pickle_fname = os.path.join(example_dir, subpath,
-                                            fname[:-5] + '_codeobj.pickle')
+                    if this_module not in doc_resolvers:
+                        continue
 
-                if os.path.exists(pickle_fname):
-                    # we have a pickle file with the objects to embed links for
-                    with open(pickle_fname, 'rb') as fid:
-                        example_code_obj = pickle.load(fid)
-                    fid.close()
-                    str_repl = {}
-                    # generate replacement strings with the links
-                    for name, cobj in example_code_obj.items():
-                        this_module = cobj['module'].split('.')[0]
-
-                        if this_module not in doc_resolvers:
-                            continue
-
+                    try:
                         link = doc_resolvers[this_module].resolve(cobj,
                                                                   full_fname)
-                        if link is not None:
-                            parts = name.split('.')
-                            name_html = period.join(orig_pattern % part
-                                                    for part in parts)
-                            str_repl[name_html] = link_pattern % (link, name_html)
-                    # do the replacement in the html file
+                    except (HTTPError, URLError) as e:
+                        print("The following error has occurred:\n")
+                        print(repr(e))
+                        continue
 
-                    # ensure greediness
-                    names = sorted(str_repl, key=len, reverse=True)
-                    expr = re.compile(r'(?<!\.>)' +  # don't follow '.' or '>'
-                                      '|'.join(re.escape(name)
-                                               for name in names))
+                    if link is not None:
+                        parts = name.split('.')
+                        name_html = period.join(orig_pattern % part
+                                                for part in parts)
+                        str_repl[name_html] = link_pattern % (link, name_html)
+                # do the replacement in the html file
 
-                    def substitute_link(match):
-                        return str_repl[match.group()]
+                # ensure greediness
+                names = sorted(str_repl, key=len, reverse=True)
+                expr = re.compile(r'(?<!\.)\b' +  # don't follow . or word
+                                  '|'.join(re.escape(name)
+                                           for name in names))
 
-                    if len(str_repl) > 0:
-                        with open(full_fname, 'rb') as fid:
-                            lines_in = fid.readlines()
-                        with open(full_fname, 'wb') as fid:
-                            for line in lines_in:
-                                line = line.decode('utf-8')
-                                line = expr.sub(substitute_link, line)
-                                fid.write(line.encode('utf-8'))
-    except HTTPError as e:
-        print("The following HTTP Error has occurred:\n")
-        print(e.code)
-    except URLError as e:
-        print("\n...\n"
-              "Warning: Embedding the documentation hyperlinks requires "
-              "internet access.\nPlease check your network connection.\n"
-              "Unable to continue embedding due to a URL Error: \n")
-        print(e.args)
+                def substitute_link(match):
+                    return str_repl[match.group()]
+
+                if len(str_repl) > 0:
+                    with open(full_fname, 'rb') as fid:
+                        lines_in = fid.readlines()
+                    with open(full_fname, 'wb') as fid:
+                        for line in lines_in:
+                            line = line.decode('utf-8')
+                            line = expr.sub(substitute_link, line)
+                            fid.write(line.encode('utf-8'))
     print('[done]')