+++ sklearn/cross_validation.py
@@ -41,6 +41,7 @@ __all__ = ['Bootstrap',
            'StratifiedShuffleSplit',
            'check_cv',
            'cross_val_score',
+           'cross_val_predict',
            'permutation_test_score',
            'train_test_split']
 
@@ -1074,6 +1075,157 @@ class StratifiedShuffleSplit(BaseShuffleSplit):
 
 
 ##############################################################################
+def cross_val_predict(estimator, X, y=None, cv=None, n_jobs=1,
+                      verbose=0, fit_params=None, pre_dispatch='2*n_jobs'):
+    """Generate cross-validated estimates for each input data point
+
+    Parameters
+    ----------
+    estimator : estimator object implementing 'fit' and 'predict'
+        The object to use to fit the data.
+
+    X : array-like
+        The data to fit. Can be, for example a list, or an array at least 2d.
+
+    y : array-like, optional, default: None
+        The target variable to try to predict in the case of
+        supervised learning.
+
+    cv : cross-validation generator or int, optional, default: None
+        A cross-validation generator to use. If int, determines
+        the number of folds in StratifiedKFold if y is binary
+        or multiclass and estimator is a classifier, or the number
+        of folds in KFold otherwise. If None, it is equivalent to cv=3.
+        This generator must include all elements in the test set exactly once.
+        Otherwise, a ValueError is raised.
+
+    n_jobs : integer, optional
+        The number of CPUs to use to do the computation. -1 means
+        'all CPUs'.
+
+    verbose : integer, optional
+        The verbosity level.
+
+    fit_params : dict, optional
+        Parameters to pass to the fit method of the estimator.
+
+    pre_dispatch : int, or string, optional
+        Controls the number of jobs that get dispatched during parallel
+        execution. Reducing this number can be useful to avoid an
+        explosion of memory consumption when more jobs get dispatched
+        than CPUs can process. This parameter can be:
+
+            - None, in which case all the jobs are immediately
+              created and spawned. Use this for lightweight and
+              fast-running jobs, to avoid delays due to on-demand
+              spawning of the jobs
+
+            - An int, giving the exact number of total jobs that are
+              spawned
+
+            - A string, giving an expression as a function of n_jobs,
+              as in '2*n_jobs'
+
+    Returns
+    -------
+    preds : ndarray
+        This is the result of calling 'predict'
+    """
+    X, y = indexable(X, y)
+
+    cv = _check_cv(cv, X, y, classifier=is_classifier(estimator))
+    # We clone the estimator to make sure that all the folds are
+    # independent, and that it is pickle-able.
+    parallel = Parallel(n_jobs=n_jobs, verbose=verbose,
+                        pre_dispatch=pre_dispatch)
+    preds_blocks = parallel(delayed(_fit_and_predict)(clone(estimator), X, y,
+                                                      train, test, verbose,
+                                                      fit_params)
+                            for train, test in cv)
+    p = np.concatenate([p for p, _ in preds_blocks])
+    locs = np.concatenate([loc for _, loc in preds_blocks])
+    if not _check_is_partition(locs, X.shape[0]):
+        raise ValueError('cross_val_predict only works for partitions')
+    preds = p.copy()
+    preds[locs] = p
+    return preds
+
+
+def _fit_and_predict(estimator, X, y, train, test, verbose, fit_params):
+    """Fit estimator and predict values for a given dataset split.
+
+    Parameters
+    ----------
+    estimator : estimator object implementing 'fit' and 'predict'
+        The object to use to fit the data.
+
+    X : array-like of shape at least 2D
+        The data to fit.
+
+    y : array-like, optional, default: None
+        The target variable to try to predict in the case of
+        supervised learning.
+
+    train : array-like, shape (n_train_samples,)
+        Indices of training samples.
+
+    test : array-like, shape (n_test_samples,)
+        Indices of test samples.
+
+    verbose : integer
+        The verbosity level.
+
+    fit_params : dict or None
+        Parameters that will be passed to ``estimator.fit``.
+
+    Returns
+    -------
+    preds : sequence
+        Result of calling 'estimator.predict'
+
+    test : array-like
+        This is the value of the test parameter
+    """
+    # Adjust length of sample weights
+    n_samples = _num_samples(X)
+    fit_params = fit_params if fit_params is not None else {}
+    fit_params = dict([(k, np.asarray(v)[train]
+                       if hasattr(v, '__len__') and len(v) == n_samples else v)
+                       for k, v in fit_params.items()])
+
+    X_train, y_train = _safe_split(estimator, X, y, train)
+    X_test, _ = _safe_split(estimator, X, y, test, train)
+
+    if y_train is None:
+        estimator.fit(X_train, **fit_params)
+    else:
+        estimator.fit(X_train, y_train, **fit_params)
+    preds = estimator.predict(X_test)
+    return preds, test
+
+
+def _check_is_partition(locs, n):
+    """Check whether locs is a reordering of the array np.arange(n)
+
+    Parameters
+    ----------
+    locs : ndarray
+        integer array to test
+    n : int
+        number of expected elements
+
+    Returns
+    -------
+    is_partition : bool
+        True iff sorted(locs) is range(n)
+    """
+    if len(locs) != n:
+        return False
+    hit = np.zeros(n, bool)
+    hit[locs] = True
+    if not np.all(hit):
+        return False
+    return True
 
 
 def cross_val_score(estimator, X, y=None, scoring=None, cv=None, n_jobs=1,
@@ -1175,10 +1327,10 @@ def _fit_and_score(estimator, X, y, scorer, train, test, verbose,
         A scorer callable object / function with signature
         ``scorer(estimator, X, y)``.
 
-    train : array-like, shape = (n_train_samples,)
+    train : array-like, shape (n_train_samples,)
         Indices of training samples.
 
-    test : array-like, shape = (n_test_samples,)
+    test : array-like, shape (n_test_samples,)
         Indices of test samples.
 
     verbose : integer
@@ -1458,7 +1610,7 @@ def permutation_test_score(estimator, X, y, cv=None,
     score : float
         The true score without permuting targets.
 
-    permutation_scores : array, shape = [n_permutations]
+    permutation_scores : array, shape (n_permutations,)
         The scores obtained for each permutations.
 
     pvalue : float
@@ -1580,8 +1732,9 @@ def train_test_split(*arrays, **options):
     if force_arrays:
         warnings.warn("The force_arrays option is deprecated and will be "
                       "removed in 0.18.", DeprecationWarning)
-        arrays = [check_array(x, 'csr', ensure_2d=False, force_all_finite=False)
-                  if x is not None else x for x in arrays]
+        arrays = [check_array(x, 'csr', ensure_2d=False,
+                              force_all_finite=False) if x is not None else x
+                  for x in arrays]
 
     if test_size is None and train_size is None:
         test_size = 0.25