+++ sklearn/dummy.py
@@ -9,12 +9,13 @@ import numpy as np
 import scipy.sparse as sp
 
 from .base import BaseEstimator, ClassifierMixin, RegressorMixin
-from .externals.six.moves import xrange
 from .utils import check_random_state
 from .utils.validation import check_array
-from sklearn.utils import deprecated
-from sklearn.utils.random import random_choice_csc
-from sklearn.utils.multiclass import class_distribution
+from .utils.validation import check_consistent_length
+from .utils import deprecated
+from .utils.random import random_choice_csc
+from .utils.stats import _weighted_percentile
+from .utils.multiclass import class_distribution
 
 
 class DummyClassifier(BaseEstimator, ClassifierMixin):
@@ -131,13 +132,13 @@ class DummyClassifier(BaseEstimator, ClassifierMixin):
          self.n_classes_,
          self.class_prior_) = class_distribution(y, sample_weight)
 
-        if self.strategy == "constant":
-            for k in range(self.n_outputs_):
-                # Checking in case of constant strategy if the constant
-                # provided by the user is in y.
-                if constant[k] not in self.classes_[k]:
-                    raise ValueError("The constant target value must be "
-                                     "present in training data")
+        if (self.strategy == "constant" and
+                any(constant[k] not in self.classes_[k]
+                    for k in range(self.n_outputs_))):
+            # Checking in case of constant strategy if the constant
+            # provided by the user is in y.
+            raise ValueError("The constant target value must be "
+                             "present in training data")
 
         if self.n_outputs_ == 1 and not self.output_2d_:
             self.n_classes_ = self.n_classes_[0]
@@ -263,7 +264,7 @@ class DummyClassifier(BaseEstimator, ClassifierMixin):
             constant = [constant]
 
         P = []
-        for k in xrange(self.n_outputs_):
+        for k in range(self.n_outputs_):
             if self.strategy == "most_frequent":
                 ind = np.ones(n_samples, dtype=int) * class_prior_[k].argmax()
                 out = np.zeros((n_samples, n_classes_[k]), dtype=np.float64)
@@ -327,6 +328,8 @@ class DummyRegressor(BaseEstimator, RegressorMixin):
 
         * "mean": always predicts the mean of the training set
         * "median": always predicts the median of the training set
+        * "quantile": always predicts a specified quantile of the training set,
+          provided with the quantile parameter.
         * "constant": always predicts a constant value that is provided by
           the user.
 
@@ -334,11 +337,16 @@ class DummyRegressor(BaseEstimator, RegressorMixin):
         The explicit constant as predicted by the "constant" strategy. This
         parameter is useful only for the "constant" strategy.
 
+    quantile : float in [0.0, 1.0]
+        The quantile to predict using the "quantile" strategy. A quantile of
+        0.5 corresponds to the median, while 0.0 to the minimum and 1.0 to the
+        maximum.
+
     Attributes
     ----------
     constant_ : float or array of shape [n_outputs]
-        Mean or median of the training targets or constant value given the by
-        the user.
+        Mean or median or quantile of the training targets or constant value
+        given by the user.
 
     n_outputs_ : int,
         Number of outputs.
@@ -347,9 +355,10 @@ class DummyRegressor(BaseEstimator, RegressorMixin):
         True if the output at fit is 2d, else false.
     """
 
-    def __init__(self, strategy="mean", constant=None):
+    def __init__(self, strategy="mean", constant=None, quantile=None):
         self.strategy = strategy
         self.constant = constant
+        self.quantile = quantile
 
     @property
     @deprecated('This will be removed in version 0.17')
@@ -358,7 +367,7 @@ class DummyRegressor(BaseEstimator, RegressorMixin):
             return self.constant_
         raise AttributeError
 
-    def fit(self, X, y):
+    def fit(self, X, y, sample_weight=None):
         """Fit the random regressor.
 
         Parameters
@@ -370,25 +379,54 @@ class DummyRegressor(BaseEstimator, RegressorMixin):
         y : array-like, shape = [n_samples] or [n_samples, n_outputs]
             Target values.
 
+        sample_weight : array-like of shape = [n_samples], optional
+            Sample weights.
+
         Returns
         -------
         self : object
             Returns self.
         """
 
-        if self.strategy not in ("mean", "median", "constant"):
-            raise ValueError("Unknown strategy type: %s, "
-                             "expected 'mean', 'median' or 'constant'"
+        if self.strategy not in ("mean", "median", "quantile", "constant"):
+            raise ValueError("Unknown strategy type: %s, expected "
+                             "'mean', 'median', 'quantile' or 'constant'"
                              % self.strategy)
 
-        y = check_array(y, accept_sparse='csr', ensure_2d=False)
-        self.output_2d_ = (y.ndim == 2)
+        y = check_array(y, ensure_2d=False)
+        if len(y) == 0:
+            raise ValueError("y must not be empty.")
+
+        self.output_2d_ = y.ndim == 2
+        if y.ndim == 1:
+            y = np.reshape(y, (-1, 1))
+        self.n_outputs_ = y.shape[1]
+
+        check_consistent_length(X, y, sample_weight)
 
         if self.strategy == "mean":
-            self.constant_ = np.reshape(np.mean(y, axis=0), (1, -1))
+            self.constant_ = np.average(y, axis=0, weights=sample_weight)
 
         elif self.strategy == "median":
-            self.constant_ = np.reshape(np.median(y, axis=0), (1, -1))
+            if sample_weight is None:
+                self.constant_ = np.median(y, axis=0)
+            else:
+                self.constant_ = [_weighted_percentile(y[:, k], sample_weight,
+                                                       percentile=50.)
+                                  for k in range(self.n_outputs_)]
+
+        elif self.strategy == "quantile":
+            if self.quantile is None or not np.isscalar(self.quantile):
+                raise ValueError("Quantile must be a scalar in the range "
+                                 "[0.0, 1.0], but got %s." % self.quantile)
+
+            percentile = self.quantile * 100.0
+            if sample_weight is None:
+                self.constant_ = np.percentile(y, axis=0, q=percentile)
+            else:
+                self.constant_ = [_weighted_percentile(y[:, k], sample_weight,
+                                                       percentile=percentile)
+                                  for k in range(self.n_outputs_)]
 
         elif self.strategy == "constant":
             if self.constant is None:
@@ -404,9 +442,9 @@ class DummyRegressor(BaseEstimator, RegressorMixin):
                     "Constant target value should have "
                     "shape (%d, 1)." % y.shape[1])
 
-            self.constant_ = np.reshape(self.constant, (1, -1))
+            self.constant_ = self.constant
 
-        self.n_outputs_ = np.size(self.constant_)  # y.shape[1] is not safe
+        self.constant_ = np.reshape(self.constant_, (1, -1))
         return self
 
     def predict(self, X):