+++ sklearn/naive_bayes.py
@@ -297,6 +297,9 @@ class GaussianNB(BaseNB):
             self.class_prior_ = np.zeros(n_classes)
             self.class_count_ = np.zeros(n_classes)
         else:
+            if X.shape[1] != self.theta_.shape[1]:
+                raise ValueError("Number of features %d does not match previous data %d." %
+                                 (X.shape[1], self.theta_.shape[1]))
             # Put epsilon back in each time
             self.sigma_[:, :] -= epsilon
 
@@ -402,6 +405,9 @@ class BaseDiscreteNB(BaseNB):
             self.class_count_ = np.zeros(n_effective_classes, dtype=np.float64)
             self.feature_count_ = np.zeros((n_effective_classes, n_features),
                                            dtype=np.float64)
+        elif n_features != self.coef_.shape[1]:
+            raise ValueError("Number of features %d does not match previous data %d."
+                             % (n_features, self.coef_.shape[-1]))
 
         Y = label_binarize(y, classes=self.classes_)
         if Y.shape[1] == 1: